Finished, Ready for Review
Added comments, finished CSS and HTML had to fix issues with the grids so made a 2 divs to fix that. The HTML had some issues because it wasn't setup correctly so trying to fix with id and classes didn't quite do it. The CSS was changed to fit the new elements added the HTML.